Se cambia la posición de la funcion create_database a la clase principal de la interfaz, para asegurarse que siempre que se inicie el programa exista la base de datos.
